Add battery error checking.
