+++ setup.c
@@ -1,4 +1,4 @@
-    #include "setup.h"
+#include "setup.h"
 
 
 void output_init(void){
@@ -38,55 +38,4 @@ void osc_init(void){
         // infinite loop, something is broken, what even is an assert()?
         while (1) {}
     }
-}
-
-static uint32_t indicator_buzzer_last_millis = 0;
-static bool buzzer_on = false;
-void indicator_buzzer_heartbeat(void){
-
-    int loop_time = millis() - indicator_buzzer_last_millis;
-
-    if(buzzer_on == false && loop_time < 750
-        && battery1_active()){
-        BUZZER_ON();
-        buzzer_on = true;
-    }
-
-    else if(buzzer_on == true && loop_time >= 750 && loop_time < 1000){
-        BUZZER_OFF();
-        buzzer_on = false;
-    }
-
-    else if(buzzer_on == false && loop_time >= 1000 && loop_time < 1250
-        && battery2_active()){
-        BUZZER_ON();
-        buzzer_on = true;
-    }
-
-    else if(buzzer_on == true && loop_time >= 1250 && loop_time < 1500){
-        BUZZER_OFF();
-        buzzer_on = false;
-    }
-
-    else if(buzzer_on == false && loop_time >= 1500 && loop_time < 1750
-        && battery2_active()){
-        BUZZER_ON();
-        buzzer_on = true;
-    }
-
-    else if(buzzer_on == true && loop_time >= 1750 && loop_time < 2000){
-        BUZZER_OFF();
-        buzzer_on = false;
-    }
-    else if(loop_time >= 2000){
-        indicator_buzzer_last_millis = millis();
-    }
-}
-
-bool battery1_active(void){
-    return (uint16_t)ADCC_GetSingleConversion(channel_BATTERY_1)*ANALOG_SCALAR > UNDERVOLTAGE_THRESHOLD;
-}
-
-bool battery2_active(void){
-    return (uint16_t)ADCC_GetSingleConversion(channel_BATTERY_2)*ANALOG_SCALAR > UNDERVOLTAGE_THRESHOLD;
 }
