+++ setup.h
@@ -8,8 +8,6 @@
 // Time between main loop code execution
 #define MAX_LOOP_TIME_DIFF_ms 1000
 
-#define UNDERVOLTAGE_THRESHOLD 5000 // in mV
-
 // LEDs
 #define RED_LED_ON() (LATC7 = 0)
 #define RED_LED_OFF() (LATC7 = 1)
@@ -34,12 +32,6 @@
 void output_init(void);
 
 void osc_init(void);
-
-void indicator_buzzer_heartbeat(void);
-
-bool battery1_active(void);
-
-bool battery2_active(void);
 
 #endif	/* ARMING_H */
